No longer dispose of servers upon Resource_invalid {safe = true}
The functionality was bugged as disposed server instances were not
reinstated after some time. Therefore after a couple of exceptions no
instances of the server was left in the pool. Once we have a mechanism
for reinstating the server instances, it might be advisable to give
server instances a timeout upon exception.